pyproject.toml
[tool.poetry]
name = "wh-imgsizer"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
packages = [
    { include = "src" },
]

[tool.poetry.dependencies]
python = "^3.8.5"
Pillow = "^7.2.0"
click = "^7.1.2"

[tool.poetry.dev-dependencies]
pytest = "^6.1.1"
flake8 = "^3.8.4"
pytest-cov = "^2.10.1"

[tool.poetry.scripts]
wh-imgsizer = 'src.scripts.run_app:main'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
